Move integration tests into separate files
