+++ .env
@@ -1,1 +1,1 @@
-DATABASE_URL=postgres://eks@localhost/eks
+DATABASE_URL=postgres://eks@127.0.0.1/eks
