+++ src/auth/mod.rs
@@ -2,14 +2,25 @@
 
 /// BSN-based identifier derivation using HKDF-SHA256.
 pub mod derive_id;
-#[cfg(feature = "dev-features")]
-pub mod dev_login;
+
 /// Session model and token utilities.
 pub mod session;
+
+/// Session storage with pluggable in-memory or Postgres backends.
+pub mod session_store;
+
 /// Postgres-backed session persistence (feature-gated).
 #[cfg(feature = "database")]
 mod session_db;
+
 /// Session middleware and request extraction.
 pub mod session_extractor;
-/// Session storage with pluggable in-memory or Postgres backends.
-pub mod session_store;
+
+#[cfg(test)]
+mod session_extractor_tests;
+
+#[cfg(feature = "dev-features")]
+pub mod dev_login;
+
+#[cfg(all(feature = "dev-features", test))]
+mod dev_login_tests;
