+++ src/core/locale_tests.rs
@@ -1,0 +1,60 @@
+use super::locale::*;
+use crate::Session;
+use axum::{
+    body::Body,
+    extract::FromRequestParts,
+    http::{Request, header},
+};
+
+#[test]
+fn converts_to_language_codes() {
+    assert_eq!(Locale::En.as_str(), "en");
+    assert_eq!(Locale::Nl.as_str(), "nl");
+}
+
+#[test]
+fn resolves_from_language_code_variants() {
+    assert_eq!(Locale::from_language_code("EN"), Some(Locale::En));
+    assert_eq!(Locale::from_language_code("nl-BE"), Some(Locale::Nl));
+    assert_eq!(Locale::from_language_code("fr"), None);
+}
+
+#[test]
+fn resolves_from_accept_language_header() {
+    let header = "nl-NL,nl;q=0.8,en;q=0.5";
+    assert_eq!(Locale::from_accept_language(header), Some(Locale::Nl));
+
+    let header = "fr-CA,fr;q=0.8,en;q=0.5";
+    assert_eq!(Locale::from_accept_language(header), None);
+}
+
+#[tokio::test]
+async fn request_locale_prefers_session() {
+    let mut request = Request::builder()
+        .uri("/")
+        .header(header::ACCEPT_LANGUAGE, "nl-NL,nl;q=0.8")
+        .body(Body::empty())
+        .unwrap();
+    request
+        .extensions_mut()
+        .insert(Session::new_test_with_locale(Locale::En));
+    let (mut parts, _body) = request.into_parts();
+
+    let locale = Locale::from_request_parts(&mut parts, &()).await.unwrap();
+
+    assert_eq!(locale, Locale::En);
+}
+
+#[tokio::test]
+async fn request_locale_falls_back_to_accept_language() {
+    let request = Request::builder()
+        .uri("/")
+        .header(header::ACCEPT_LANGUAGE, "nl-NL,nl;q=0.8")
+        .body(Body::empty())
+        .unwrap();
+    let (mut parts, _body) = request.into_parts();
+
+    let locale = Locale::from_request_parts(&mut parts, &()).await.unwrap();
+
+    assert_eq!(locale, Locale::Nl);
+}
