+++ src/error/mod.rs
@@ -1,5 +1,11 @@
 mod app_error;
 mod response;
 
+#[cfg(test)]
+mod app_error_tests;
+
+#[cfg(test)]
+mod response_tests;
+
 pub use app_error::{AppError, AppResponse};
 pub use response::{ErrorResponse, render_error_pages};
